Authorization with Gates & Policies .. Done!

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Traits\CanLoadRelationShips;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -20,6 +21,7 @@ class AttendeeController extends Controller
      */
     public function index(Event $event)
     {
+        Gate::authorize('viewAny', Attendee::class);
         $attendees = $event->attendees()->latest();
         return AttendeeResouce::collection($this->loadRelationships($attendees)->paginate());
     }
@@ -29,6 +31,7 @@ class AttendeeController extends Controller
      */
     public function store(Request $request, Event $event)
     {
+        Gate::authorize('create', Attendee::class);
         $validatedData = $request->validate([
             'user_id' => 'required',
         ]);
@@ -42,6 +45,8 @@ class AttendeeController extends Controller
      */
     public function show(Event $event, Attendee $attendee)
     {
+        Gate::authorize('view', $attendee);
+
         return new AttendeeResouce($this->loadRelationships($attendee));
     }
 
@@ -51,6 +56,9 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        Gate::authorize('delete', $attendee);
+
+        // Gate::authorize('delete-attendee', [$event, $attendee]);
         $attendee->deleteOrFail();
         return response(status:204);
     }
